backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
deprecation==2.1.0
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
gotrue==2.4.4
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
packaging==24.1
pillow==10.3.0
postgrest==0.16.8
pydantic==2.7.4
pydantic_core==2.18.4
python-dateutil==2.9.0.post0
realtime==1.0.6
reportlab==4.2.0
requests==2.32.3
resend==2.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.6
StrEnum==0.4.15
supabase==2.5.1
supafunc==0.4.6
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websockets==12.0
